+++ src/test/java/io/github/astrapi69/check/CheckTest.java
@@ -31,7 +31,7 @@ import java.util.Map;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
-import io.github.astrapi69.collections.array.ArrayFactory;
+import io.github.astrapi69.collection.array.ArrayFactory;
 import io.github.astrapi69.test.object.Person;
 
 /**