+++ src/test/java/io/github/astrapi69/lang/model/FieldModelTest.java
@@ -31,7 +31,7 @@ import nl.jqno.equalsverifier.EqualsVerifier;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
-import io.github.astrapi69.collections.list.ListFactory;
+import io.github.astrapi69.collection.list.ListFactory;
 
 /**
  * The unit test class for the class {@link FieldModel}.