+++ src/test/java/io/github/astrapi69/lang/model/MethodModelTest.java
@@ -32,8 +32,8 @@ import nl.jqno.equalsverifier.EqualsVerifier;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
-import io.github.astrapi69.collections.list.ListFactory;
-import io.github.astrapi69.collections.map.MapFactory;
+import io.github.astrapi69.collection.list.ListFactory;
+import io.github.astrapi69.collection.map.MapFactory;
 
 /**
  * The unit test class for the class {@link MethodModel}.